+++ lib/ui/components/cart/shop_item_list.dart
@@ -10,7 +10,7 @@ class ShopItemList extends StatelessWidget {
   final Function(Item)? onLongPress;
   final Function(Item) onAddToCart;
   final Function(Item) showVariants;
-  final int qtyOnCart;
+  final double qtyOnCart;
 
   const ShopItemList({
     required this.item,
